requirements.txt
annotated-types==0.7.0
anyio==4.5.2
audeer==2.2.0
audformat==1.1.4
audinterface==1.2.1
audiofile==1.5.0
audmath==1.4.1
audobject==0.7.11
audresample==1.3.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iso-639==0.4.5
iso3166==2.1.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
neurokit2==0.2.10
numba==0.58.1
numpy==1.24.4
openai-whisper==20240930
opensmile==2.4.2
oyaml==1.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip-autoremove==0.10.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
